2002-11-11  Michael Koch <konqueror@gmx.de>
	* java/nio/channels/AlreadyConnectedException.java:
	Removed unneeded import.
	(AlreadyConnectedException): Added documentation.
	* java/nio/channels/ByteChannel.java: Wrapped.
	* java/nio/channels/Channel.java: Documentation added.
	* java/nio/channels/DatagramChannel.java: Reindented.
	* java/nio/channels/FileChannel.java: Reindented.
	* java/nio/channels/GatheringByteChannel.java: Documentation added.
	* java/nio/channels/InterruptibleChannel.java: Documentation added.
	* java/nio/channels/ReadableByteChannel.java: Documentation added.
	* java/nio/channels/ScatteringByteChannel.java: Documentation added.
	* java/nio/channels/WritableByteChannel.java: Documentation added.
	* java/nio/channels/spi/AbstractChannel.java: Reindented.
	* java/nio/channels/spi/AbstractSelectableChannel.java:
	(blockingLock): Must be final.
	(configureBlocking): Must be final.
	(implCloseChannel): Must be final.
	(isBlocking): Must be final.
	(isRegistered): Must be final.
	(provider): Must be final.
	(register): Must be final.
	* java/nio/channels/spi/AbstractSelectionKey.java:
	Reindented.
	(cancel): Must be final.
	(isValid): Must be final.

+++ java/nio/channels/GatheringByteChannel.java
@@ -37,12 +37,43 @@ exception statement from your version. */
 
 package java.nio.channels;
 
+import java.nio.ByteBuffer;
+import java.io.IOException;
 
-import java.nio.*;
-import java.io.*;
-
-public interface GatheringByteChannel extends WritableByteChannel
+public interface GatheringByteChannel
+  extends WritableByteChannel
 {
-    public long write(ByteBuffer[] srcs, int offset, int length) throws IOException;
-    public int write(ByteBuffer[] srcs) throws IOException;
+  /**
+   * Writes a sequence of bytes to this channel from a subsequence of
+   * the given buffers
+   *
+   * @exception AsynchronousCloseException If another thread closes this
+   * channel while the write operation is in progress
+   * @exception ClosedByInterruptException If another thread interrupts the
+   * current thread while the write operation is in progress, thereby closing
+   * the channel and setting the current thread's interrupt status
+   * @exception ClosedChannelException If this channel is closed
+   * @exception IndexOutOfBoundsException If the preconditions on the offset
+   * and length parameters do not hold
+   * @exception IOException If an error occurs
+   * @exception NonWritableChannelException If this channel was not opened for
+   * writing
+   */
+  public long write(ByteBuffer[] srcs, int offset, int length)
+    throws IOException;
+  
+  /**
+   * Writes a sequence of bytes to this channel from the given buffers
+   *
+   * @exception AsynchronousCloseException If another thread closes this
+   * channel while the write operation is in progress
+   * @exception ClosedByInterruptException If another thread interrupts the
+   * current thread while the write operation is in progress, thereby closing
+   * the channel and setting the current thread's interrupt status
+   * @exception ClosedChannelException If this channel is closed
+   * @exception IOException If an error occurs
+   * @exception NonWritableChannelException If this channel was not opened for
+   * writing
+   */
+  public long write(ByteBuffer[] srcs) throws IOException;
 }
